requirements-dev.txt
qiskit-optimization
qiskit-ibmq-provider>=0.18
qiskit-nature
pyyaml
requests
pylint==2.13.8
black==22.3.0
stestr>=2.0.0
ddt>=1.2.0,!=1.4.0,!=1.4.3
mypy>=0.780
mypy-extensions>=0.4.3
oct2py
pandas
types-requests
types-PyYAML
